更新 Makefile 和 Makefile_mb，添加汇编和链接器标志以消除警告；在多个源文件中修复指针类型转换以增强代码安全性

+++ PocketNES/src/cache.c
@@ -163,15 +163,15 @@ void get_rom_map()
 			#if PRG_BANK_SIZE == 8
 			*((u8**)&readmem_6)=(u8*)sram_R;
 			*((u8**)&writemem_6)=(u8*)sram_W_func;
-			*((u8**)&memmap_6)=(u8*)(NES_RAM-0x5800);
+			*((u8**)&memmap_6)=(u8*)((u32)NES_RAM-0x5800);
 			#endif
 			#if PRG_BANK_SIZE == 4
 			*((u8**)&readmem_6)=(u8*)sram_R;
 			*((u8**)&readmem_7)=(u8*)sram_R;
 			*((u8**)&writemem_6)=(u8*)sram_W_func;
 			*((u8**)&writemem_7)=(u8*)sram_W_func;
-			*((u8**)&memmap_6)=(u8*)(NES_RAM-0x5800);
-			*((u8**)&memmap_7)=(u8*)(NES_RAM-0x5800);
+			*((u8**)&memmap_6)=(u8*)((u32)NES_RAM-0x5800);
+			*((u8**)&memmap_7)=(u8*)((u32)NES_RAM-0x5800);
 			#endif
 		}
 		else
